urbanopt_climate_sweep now runs both scenarios, also added scenario report
I had planned on adding argument to run one or the other scenario, but the way code is setup it was easier to have it create and run the scenarios when each project is first made. I also enabled post processing and feature reporting and added scenario reporting to compare two scenarios against each other.